+++ client/packages/portal-pages/src/index.ts
@@ -1,2 +1,3 @@
 export * from './pages/home-page/HomePage';
 export * from './pages/phase-page/PhasePage';
+export * from './pages/home-page/LoadingWorkSurfacesTransition';
