+++ client/packages/portal-pages/src/pages/home-page/LoadingWorkSurfacesTransition.tsx
